feat: implement `throttle` function
Sends elements to the returned channel limiting the throughput to specific
number of elements (evenly spaced) per time unit. Note that the element's
`receive()` time is included in the resulting throughput. For instance having
`throttle(1, 1.second)` and `receive()` taking `Xms` means that resulting
channel will receive elements every `1s + Xms` time.
Throttling is not applied to the empty source.

Examples:

  Source.empty[Int].throttle(1, 1.second).toList       // List() returned without throttling
  Source.fromValues(1, 2).throttle(1, 1.second).toList // List(1, 2) returned after 2 seconds

Note that implementation relies on `Thread.sleep` that is according to
[1] project Loom compatible.

[1] https://softwaremill.com/what-is-blocking-in-loom/