feat: implement `last` and `lastOption` operators
The `lastOption` operator returns the last element in `Source` wrapped in
`Some` or `None` in case when source is empty. Note that this is a terminal
operation for source e.g.:

  Source.empty[Int].lastOption()  // None
  val s = Source.fromValues(1, 2)
  s.lastOption()                  // Some(2)
  s.receive()                     // ChannelClosed.Done

The `last` operator returns the last element in `Source` or throws
`NoSuchElementException` in case when it is empty. In case when `receive()`
fails then `ChannelClosedException.Error` exception is thrown. It is also
a terminal operation e.g.:

  Source.empty[Int].last()        // throws NoSuchElementException("cannot obtain last element from an empty source")
  val s = Source.fromValues(1, 2)
  s.last()                        // 2
  s.receive()                     // ChannelClosed.Done

Note that ChannelClosedException.Error was improved to contain `cause`
exception (if available).